Fold restore prep into restore builder
