+++ deploy/snapshot/workload/restore_pod.go
@@ -26,3 +26,85 @@ func NewRestorePod(pod *corev1.Pod, opts RestorePodOptions) *corev1.Pod {
 	pod.Spec.RestartPolicy = corev1.RestartPolicyNever
 	return pod
 }
+
+func PrepareRestorePodSpec(
+	podSpec *corev1.PodSpec,
+	container *corev1.Container,
+	pvcName string,
+	basePath string,
+	seccompProfile string,
+	placeholder bool,
+) {
+	injectLocalhostSeccompProfile(podSpec, seccompProfile)
+	if pvcName != "" {
+		injectCheckpointVolume(podSpec, pvcName)
+	}
+	if basePath != "" {
+		injectCheckpointVolumeMount(container, basePath)
+	}
+	injectRestoreTUN(podSpec, container)
+	if placeholder {
+		container.Command = []string{"sleep", "infinity"}
+		container.Args = nil
+	}
+}
+
+func injectCheckpointVolume(podSpec *corev1.PodSpec, pvcName string) {
+	for _, volume := range podSpec.Volumes {
+		if volume.Name == CheckpointVolumeName {
+			return
+		}
+	}
+
+	podSpec.Volumes = append(podSpec.Volumes, corev1.Volume{
+		Name: CheckpointVolumeName,
+		VolumeSource: corev1.VolumeSource{
+			PersistentVolumeClaim: &corev1.PersistentVolumeClaimVolumeSource{
+				ClaimName: pvcName,
+			},
+		},
+	})
+}
+
+func injectCheckpointVolumeMount(container *corev1.Container, basePath string) {
+	for _, mount := range container.VolumeMounts {
+		if mount.Name == CheckpointVolumeName {
+			return
+		}
+	}
+
+	container.VolumeMounts = append(container.VolumeMounts, corev1.VolumeMount{
+		Name:      CheckpointVolumeName,
+		MountPath: basePath,
+	})
+}
+
+func injectRestoreTUN(podSpec *corev1.PodSpec, container *corev1.Container) {
+	charDevice := corev1.HostPathCharDev
+
+	for _, volume := range podSpec.Volumes {
+		if volume.Name == RestoreTUNVolumeName {
+			goto mount
+		}
+	}
+	podSpec.Volumes = append(podSpec.Volumes, corev1.Volume{
+		Name: RestoreTUNVolumeName,
+		VolumeSource: corev1.VolumeSource{
+			HostPath: &corev1.HostPathVolumeSource{
+				Path: "/dev/net/tun",
+				Type: &charDevice,
+			},
+		},
+	})
+
+mount:
+	for _, mount := range container.VolumeMounts {
+		if mount.Name == RestoreTUNVolumeName {
+			return
+		}
+	}
+	container.VolumeMounts = append(container.VolumeMounts, corev1.VolumeMount{
+		Name:      RestoreTUNVolumeName,
+		MountPath: "/dev/net/tun",
+	})
+}
